Fix autocomplete test
The test fail because gotoline now reacts in a different way when supplying out of range arguments: the first block of document is returned. The new values passed to this function are really what we want to test (this is a left over from the time where line and column were 1 based).